Split mine model class file
