+++ python-backend/app/api/parties/party/models/party.py
@@ -8,7 +8,7 @@ from sqlalchemy.ext.hybrid import hybrid_property
 from sqlalchemy.orm import validates
 from app.extensions import db
 
-from ....mines.mine.models.mine import MineIdentity
+from ....mines.mine.models.mine_identity import MineIdentity
 from ....utils.models_mixins import AuditMixin, Base
 from ....constants import PARTY_STATUS_CODE
