+++ python-backend/app/commands.py
@@ -10,7 +10,9 @@ from .api.mines.location.models.mine_location import MineLocation
 from .api.mines.region.models.region import MineRegionCode
 from .api.constants import PERMIT_STATUS_CODE, MINE_OPERATION_STATUS, MINE_OPERATION_STATUS_REASON, \
 MINE_OPERATION_STATUS_SUB_REASON, MINE_REGION_OPTIONS
-from .api.mines.mine.models.mine import MineIdentity, MineDetail, MineralTenureXref
+from .api.mines.mine.models.mine_identity import MineIdentity
+from .api.mines.mine.models.mine_detail import MineDetail
+from .api.mines.mine.models.mineral_tenure_xref import MineralTenureXref
 from .api.parties.party.models.party import Party, PartyTypeCode
 from .api.permits.permit.models.permit import Permit, PermitStatusCode
 from .api.permits.permittee.models.permittee import Permittee
